Step 1: upgrade SBT plugins in project/plugins.sbt
It is possible that there may be a number of HMRC SBT plugins that require updating to Play 2.5 compatible versions. Ensure you are using the latest. However, the most important thing is that you upgrade Typesafe's sbt-plugin to version 2.5.8.

It is most likely that you will need to update:

1. addSbtPlugin("uk.gov.hmrc" % "sbt-distributables" % "1.0.0")

2. addSbtPlugin("com.typesafe.play" % "sbt-plugin" % "2.5.8")

NOTE: This change will break your project. If you are using IntelliJ with auto-import of SBT projects, you will probably see a project import error.