Step 3: VERY IMPORTANT - update dependencies in project/FrontendBuild.scala
Update your HMRC dependencies to the latest versions which are designed to support Play 2.5. See https://confluence.tools.tax.service.gov.uk/pages/viewpage.action?spaceKey=DTRG&title=Play+2.5.8+Upgrade+Guide for a full list of upgraded libraries. Upgrade any and all libraries listed in that Confluence page that you utilise.

You may also need to update your versions of 3rd party Play-dependent libraries, such as play-test and scalatest+play amongst others. It is important for YOU to verify that all of your dependencies are Play 2.5 compatible.

If you are seeing eviction warnings for Play 2.3, you MUST pay attention as they are almost certain to cause compilation failures (if you are lucky) or, more often, unexpected runtime errors. Not all evictions are a problem but ones relating to core play libraries are problematic.

It is quite possible that you depend on an HMRC library (possibly created by your team and only used by yourselves) that depends on Play but which has NOT been upgraded yet. In such cases, it is YOUR responsibility to upgrade the library.

After completing this step, it is almost certain that your project will still not compile.