Step 5: fix the inevitable compilation errors
It is almost 100% certain that you will now have a non-compiling project. The volume of compilation errors will obviously depend on the size of your project and its usage of the Play APIs. Attempts have been made wherever possible in supporting HMRC libraries to mitigate the effects of this.

For frontends, the volume of compilation errors introduced as a result of upgrading to Play 2.5 is liable to be pretty high. However, the majority of these will probably be the result of changes to the i18n API. Any use of `@Messages` in Twirl templates now require an implicit `play.api.i18n.Messages` instance. This messages API encapsulates access to Lang, so you might also be able to remove a lot of implicit references to Lang which are no longer required.

For a blank frontend created using init-service, the following Twirl templates needed to have an implicit messages parameter added:

1. govuk_wrapper.scala.html
2. main_template.scala.html
3. error_template.scala.html
4. hello_world.scala.html

As a result of introducing the implicit `Messages` parameter to the Twirl templates, you get a cascade of compilation failures as wherever they are referenced now also needs a reference to `Messages`. Obviously, Play recommend that you use DI to obtain this reference. However, if you are unwilling (e.g. because you do not want to migrate all your object-based controllers yet) or unable (e.g. because you cannot dependency-inject an application global) to use DI, there is a workaround.

A. `import play.api.i18n.Messages.Implicits._`
B. `import play.api.Play.current` (deprecated, but still works)
C. You may need to Explicitly pass the implicit messages parameter where required using `applicationMessages`, which is provided by `play.api.i18n.Messages.Implicits`

The first place we need to use this workaround is for the error handler methods in `frontendGlobal`. You will also almost certainly need to use this workaround in your controllers (or migrate them to use DI).

Additionally, if you have filters, they now require you to provide an Akka Materializer. The play-filters library provides a MicroserviceFilterSupport mixin trait which does this for you. Simply mix it in to your filter object.

Finally, support for `onLoadConfig` has been entirely removed from `GlobalSettings` in Play 2.5. Fortunately, we weren't actually using it but, if your specific frontend does, then you will need to find an alternative technique to achieve your aim in doing so. See https://www.playframework.com/documentation/2.5.x/GlobalSettings#Scala. More likely is that the init-service script will have placed a pointless override of `onLoadConfig` which is still sitting there unused and that now needs to be removed, if present, from your `frontendGlobal`.